requirements.txt
matplotlib
pyelftools
rich
